Create checkstake.sh
This bash script sends email notifications when QTUM wallet "stake" value changes.  It sends both a short email-to-text and a full email with more info.  It will also notify if the QTUM wallet is not "unlocked", which can happen after a reboot.  

When a stake change is detected, the script waits until 6 confirmations (blocks) have been added on top of the staked block before logging the "stake change" and sending out the notifications.

If the stake value reverts to the previous value before 6 confirmations, the script logs it as a "stake orphan" and sends no notification.